+++ server/src-lib/Hasura/RQL/DDL/Schema/Cache.hs
@@ -28,6 +28,7 @@ import qualified Data.HashMap.Strict.Extended             as M
 import qualified Data.HashMap.Strict.InsOrd               as OMap
 import qualified Data.HashSet                             as HS
 import qualified Data.HashSet.InsOrd                      as HSIns
+import qualified Data.Set                                 as S
 import qualified Language.GraphQL.Draft.Syntax            as G
 
 import           Control.Arrow.Extended
@@ -168,6 +169,12 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
     duplicateVariables :: EndpointMetadata a -> Bool
     duplicateVariables m = any ((>1) . length) $ group $ sort $ catMaybes $ splitPath Just (const Nothing) (_ceUrl m)
 
+    endpointObjId :: EndpointMetadata q -> MetadataObjId
+    endpointObjId md = MOEndpoint (_ceName md)
+
+    endpointObject :: EndpointMetadata q -> MetadataObject
+    endpointObject md = MetadataObject (endpointObjId md) (toJSON $ OMap.lookup (_ceName md) $ _metaRestEndpoints metadata)
+
     --  Cases of urls that generate invalid segments:
     -- * "" -> Error: "Empty URL"
     -- * "/asdf" -> Error: "Leading slash not allowed"
@@ -175,18 +182,22 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
     -- * "asdf//qwer" -> Error: "Double Slash not allowed"
     -- * "asdf/:/qwer" -> Error: "Variables must be named"
     -- * "asdf/:x/qwer/:x" -> Error: "Duplicate Variable: x"
-    invalidSegments m = any (`elem` ["",":"]) (splitPath id id (_ceUrl m))
+    hasInvalidSegments :: EndpointMetadata query -> Bool
+    hasInvalidSegments m = any (`elem` ["",":"]) (splitPath id id (_ceUrl m))
 
-  bindA -< onJust (nonEmpty $ filter duplicateVariables (M.elems $ _boEndpoints resolvedOutputs)) $ \md ->
-      throw400 BadRequest $ "Duplicate variables found in endpoint paths: " <> commaSeparated (_ceUrl <$> md)
+    ceUrlTxt               = toTxt . _ceUrl
 
-  bindA -< onJust (nonEmpty $ filter invalidSegments (M.elems $ _boEndpoints resolvedOutputs)) $ \md ->
-      throw400 BadRequest $ "Empty segments or unnamed variables are not allowed: " <> commaSeparated (_ceUrl <$> md)
+    endpoints              = buildEndpointsTrie (M.elems $ _boEndpoints resolvedOutputs)
 
-  let endpoints = buildEndpointsTrie (M.elems $ _boEndpoints resolvedOutputs)
+    duplicateF md          = DuplicateRestVariables ("Duplicate variables found in endpoint path " <> (ceUrlTxt md)) (endpointObject md)
+    duplicateRestVariables = map duplicateF $ filter duplicateVariables (M.elems $ _boEndpoints resolvedOutputs)
 
-  bindA -< onJust (nonEmpty $ ambiguousPaths endpoints) $ \ambPaths ->
-      throw409 $ "Ambiguous URL paths in endpoints: " <> commaSeparated (renderPath <$> ambPaths)
+    invalidF md            = InvalidRestSegments ("Empty segments or unnamed variables are not allowed: " <> (ceUrlTxt md)) (endpointObject md)
+    invalidRestSegments    = map invalidF $ filter hasInvalidSegments (M.elems $ _boEndpoints resolvedOutputs)
+
+    ambiguousF' ep         = MetadataObject (endpointObjId ep) (toJSON ep)
+    ambiguousF mds         = AmbiguousRestEndpoints ("Ambiguous URL paths: " <> commaSeparated (map _ceUrl mds)) (map ambiguousF' mds)
+    ambiguousRestEndpoints = map (ambiguousF . S.elems . snd) $ ambiguousPathsGrouped endpoints
 
   returnA -< SchemaCache
     { scSources = _boSources resolvedOutputs
@@ -210,6 +221,9 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
         <> dependencyInconsistentObjects
         <> toList gqlSchemaInconsistentObjects
         <> toList relaySchemaInconsistentObjects
+        <> duplicateRestVariables
+        <> invalidRestSegments
+        <> ambiguousRestEndpoints
     , scApiLimits = _boApiLimits resolvedOutputs
     , scMetricsConfig = _boMetricsConfig resolvedOutputs
     }
@@ -264,8 +278,9 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
          , DBFunctionsMetadata b
          , RemoteSchemaMap
          , Inc.Dependency InvalidationKeys
+         , InheritedRoles
          ) `arr` BackendSourceInfo
-    buildSource = proc (sourceMetadata, sourceConfig, dbTables, dbFunctions, remoteSchemaMap, invalidationKeys) -> do
+    buildSource = proc (sourceMetadata, sourceConfig, dbTables, dbFunctions, remoteSchemaMap, invalidationKeys, inheritedRoles) -> do
       let SourceMetadata source tables functions _ = sourceMetadata
           tablesMetadata = OMap.elems tables
           (tableInputs, nonColumnInputs, permissions) = unzip3 $ map mkTableInputs tablesMetadata
@@ -292,7 +307,9 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
       tableCache <-
         (| Inc.keyed (\_ ((tableCoreInfo, permissionInputs), (_, eventTriggerConfs)) -> do
              let tableFields = _tciFieldInfoMap tableCoreInfo
-             permissionInfos <- buildTablePermissions -< (source, tableCoreInfosDep, tableFields, permissionInputs)
+             permissionInfos <-
+                buildTablePermissions
+                -< (source, tableCoreInfosDep, tableFields, permissionInputs, inheritedRoles)
              eventTriggerInfos <- buildTableEventTriggers -< (source, sourceConfig, tableCoreInfo, eventTriggerConfs, metadataInvalidationKey)
              returnA -< TableInfo tableCoreInfo permissionInfos eventTriggerInfos
             )
@@ -330,15 +347,16 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
       => ( Inc.Dependency InvalidationKeys
          , HashMap RemoteSchemaName RemoteSchemaCtx
          , SourceMetadata b
+         , InheritedRoles
          ) `arr` Maybe (BackendSourceInfo, DMap.DMap BackendTag ScalarSet)
-    buildSourceOutput = proc (invalidationKeys, remoteSchemaCtxMap, sourceMetadata) -> do
+    buildSourceOutput = proc (invalidationKeys, remoteSchemaCtxMap, sourceMetadata, inheritedRoles) -> do
       let sourceInvalidationsKeys = Inc.selectD #_ikSources invalidationKeys
       maybeResolvedSource <- resolveSourceIfNeeded -< (sourceInvalidationsKeys, sourceMetadata)
       case maybeResolvedSource of
         Nothing -> returnA -< Nothing
         Just (ResolvedSource sourceConfig tablesMeta functionsMeta scalars) -> do
           so <- buildSource -< ( sourceMetadata, sourceConfig, tablesMeta, functionsMeta
-                               , remoteSchemaCtxMap, invalidationKeys
+                               , remoteSchemaCtxMap, invalidationKeys, inheritedRoles
                                )
           returnA -< Just (so, DMap.singleton (backendTag @b) $ ScalarSet scalars)
 
@@ -351,7 +369,16 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
       => (Metadata, Inc.Dependency InvalidationKeys) `arr` BuildOutputs
     buildAndCollectInfo = proc (metadata, invalidationKeys) -> do
       let Metadata sources remoteSchemas collections allowlists
-            customTypes actions cronTriggers endpoints apiLimits metricsConfig = metadata
+            customTypes actions cronTriggers endpoints apiLimits metricsConfig inheritedRoles = metadata
+          actionRoles = map _apmRole . _amPermissions =<< OMap.elems actions
+          remoteSchemaRoles = map _rspmRole . _rsmPermissions =<< OMap.elems remoteSchemas
+          sourceRoles =
+            HS.fromList $ concat $
+            OMap.elems sources >>= \(BackendSourceMetadata (SourceMetadata _ tables _functions _) ) -> do
+               table <- OMap.elems tables
+               pure ( OMap.keys (_tmInsertPermissions table) <> OMap.keys (_tmSelectPermissions table)
+                    <> OMap.keys (_tmUpdatePermissions table) <> OMap.keys (_tmDeletePermissions table))
+
           remoteSchemaPermissions =
             let remoteSchemaPermsList = OMap.toList $ _rsmPermissions <$> remoteSchemas
             in concat $ flip map remoteSchemaPermsList $
@@ -359,6 +386,22 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
                     flip map remoteSchemaPerms $ \(RemoteSchemaPermissionMetadata role defn comment) ->
                      AddRemoteSchemaPermissions remoteSchemaName role defn comment
                  )
+          nonInheritedRoles = sourceRoles <> HS.fromList (actionRoles <> remoteSchemaRoles)
+
+
+      let commonInheritedRoles = HS.intersection (HS.fromList (OMap.keys inheritedRoles)) nonInheritedRoles
+
+
+      bindA -< do
+        unless (HS.null commonInheritedRoles) $ do
+          throw400 AlreadyExists $
+            "role " <> commaSeparated (map toTxt $ toList commonInheritedRoles) <> " already exists"
+        for_ (toList inheritedRoles) $ \(AddInheritedRole _ roleSet) ->
+          for_ roleSet $ \role -> do
+            unless (role `elem` nonInheritedRoles) $
+              throw400 NotFound $ role <<> " not found. An inherited role can only be created out of existing roles"
+            when (role `OMap.member` inheritedRoles) $
+              throw400 ConstraintError $  role <<> " is an inherited role. An inherited role can only be created out of non-inherited roles"
 
       -- remote schemas
       let remoteSchemaInvalidationKeys = Inc.selectD #_ikRemoteSchemas invalidationKeys
@@ -382,8 +425,8 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
       sourcesOutput <-
         (| Inc.keyed (\_ (BackendSourceMetadata (sourceMetadata :: SourceMetadata b)) ->
              case backendTag @b of
-                PostgresTag -> buildSourceOutput @arr @m -< (invalidationKeys, remoteSchemaCtxMap, sourceMetadata :: SourceMetadata 'Postgres)
-                MSSQLTag    -> buildSourceOutput @arr @m -< (invalidationKeys, remoteSchemaCtxMap, sourceMetadata :: SourceMetadata 'MSSQL)
+                PostgresTag -> buildSourceOutput @arr @m -< (invalidationKeys, remoteSchemaCtxMap, sourceMetadata :: SourceMetadata 'Postgres, inheritedRoles)
+                MSSQLTag    -> buildSourceOutput @arr @m -< (invalidationKeys, remoteSchemaCtxMap, sourceMetadata :: SourceMetadata 'MSSQL, inheritedRoles)
            )
         |) (M.fromList $ OMap.toList sources)
         >-> (\infos -> M.catMaybes infos >- returnA)
@@ -424,6 +467,8 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
 
       cronTriggersMap <- buildCronTriggers -< ((), OMap.elems cronTriggers)
 
+      let inheritedRolesCache = OMap.toHashMap $ fmap _adrRoleSet inheritedRoles
+
       returnA -< BuildOutputs
         { _boSources = M.map fst sourcesOutput
         , _boActions = actionCache
@@ -431,6 +476,7 @@ buildSchemaCacheRule env = proc (metadata, invalidationKeys) -> do
         , _boAllowlist = allowList
         , _boCustomTypes = annotatedCustomTypes
         , _boCronTriggers = cronTriggersMap
+        , _boInheritedRoles = inheritedRolesCache
         , _boEndpoints = resolvedEndpoints
         , _boApiLimits = apiLimits
         , _boMetricsConfig = metricsConfig