+++ server/src-lib/Hasura/RQL/Types/Metadata/Object.hs
@@ -52,6 +52,7 @@ data MetadataObjId
   | MOAction !ActionName
   | MOActionPermission !ActionName !RoleName
   | MOCronTrigger !TriggerName
+  | MOInheritedRole !RoleName
   | MOEndpoint !EndpointName
 $(makePrisms ''MetadataObjId)
 
@@ -65,6 +66,7 @@ instance Hashable MetadataObjId where
     MOAction actionName                    -> hashWithSalt salt actionName
     MOActionPermission actionName roleName -> hashWithSalt salt (actionName, roleName)
     MOCronTrigger triggerName              -> hashWithSalt salt triggerName
+    MOInheritedRole roleName               -> hashWithSalt salt roleName
     MOEndpoint endpoint                    -> hashWithSalt salt endpoint
 
 instance Eq MetadataObjId where
@@ -75,6 +77,7 @@ instance Eq MetadataObjId where
   MOCustomTypes == MOCustomTypes                             = True
   (MOActionPermission an1 r1) == (MOActionPermission an2 r2) = an1 == an2 && r1 == r2
   (MOCronTrigger trn1) == (MOCronTrigger trn2)               = trn1 == trn2
+  (MOInheritedRole rn1) == (MOInheritedRole rn2)             = rn1 == rn2
   _ == _                                                     = False
 
 moiTypeName :: MetadataObjId -> Text
@@ -96,6 +99,7 @@ moiTypeName = \case
   MOCustomTypes -> "custom_types"
   MOAction _ -> "action"
   MOActionPermission _ _ -> "action_permission"
+  MOInheritedRole _      -> "inherited_role"
   MOEndpoint _ -> "endpoint"
 
 moiName :: MetadataObjId -> Text
@@ -122,6 +126,7 @@ moiName objectId = moiTypeName objectId <> " " <> case objectId of
   MOCustomTypes -> "custom_types"
   MOAction name -> toTxt name
   MOActionPermission name roleName -> toTxt roleName <> " permission in " <> toTxt name
+  MOInheritedRole inheritedRoleName -> "inherited role " <> toTxt inheritedRoleName
   MOEndpoint name -> toTxt name
 
 data MetadataObject
@@ -135,6 +140,9 @@ data InconsistentMetadata
   = InconsistentObject !Text !MetadataObject
   | ConflictingObjects !Text ![MetadataObject]
   | DuplicateObjects !MetadataObjId ![Value]
+  | DuplicateRestVariables !Text !MetadataObject
+  | InvalidRestSegments !Text !MetadataObject
+  | AmbiguousRestEndpoints !Text ![MetadataObject]
   deriving (Eq)
 $(makePrisms ''InconsistentMetadata)
 
@@ -147,12 +155,18 @@ imObjectIds = \case
   InconsistentObject _ metadata  -> [_moId metadata]
   ConflictingObjects _ metadatas -> map _moId metadatas
   DuplicateObjects objectId _    -> [objectId]
+  DuplicateRestVariables _ md    -> [_moId md]
+  InvalidRestSegments _ md       -> [_moId md]
+  AmbiguousRestEndpoints _ mds   -> take 1 $ map _moId mds -- TODO: Take 1 is a workaround to ensure that conflicts are not reported multiple times per endpoint.
 
 imReason :: InconsistentMetadata -> Text
 imReason = \case
-  InconsistentObject reason _ -> reason
-  ConflictingObjects reason _ -> reason
-  DuplicateObjects objectId _ -> "multiple definitions for " <> moiName objectId
+  InconsistentObject reason _     -> reason
+  ConflictingObjects reason _     -> reason
+  DuplicateObjects objectId _     -> "multiple definitions for " <> moiName objectId
+  DuplicateRestVariables reason _ -> reason
+  InvalidRestSegments reason _    -> reason
+  AmbiguousRestEndpoints reason _ -> reason
 
 -- | Builds a map from each unique metadata object id to the inconsistencies associated with it.
 -- Note that a single inconsistency can involve multiple metadata objects, so the same inconsistency
@@ -172,6 +186,10 @@ instance ToJSON InconsistentMetadata where
         DuplicateObjects objectId definitions ->
           [ "type" .= String (moiTypeName objectId)
           , "definitions" .= definitions ]
+
+        DuplicateRestVariables _ md  -> metadataObjectFields md
+        InvalidRestSegments _ md     -> metadataObjectFields md
+        AmbiguousRestEndpoints _ mds -> [ "conflicts" .= map _moDefinition mds ]
 
       metadataObjectFields (MetadataObject objectId definition) =
         [ "type" .= String (moiTypeName objectId)